reorganized code to make it easier for updates
